requirements.txt
fastapi
uvicorn
pytesseract
openai
python-multipart
Pillow